feat : integration users home page with real api
