+++ src/app/(main)/(user)/layout.tsx
@@ -1,0 +1,14 @@
+import SidebarContent from "@/app/components/SidebarContent";
+
+export default function UserLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <SidebarContent activeView="bookings" mockBookings={[]} />
+      {children}
+    </>
+  );
+}
