+++ src/store/booking-store.ts
@@ -1,15 +1,19 @@
 import { create } from "zustand";
 import axiosInstance from "@/lib/axiosInstance";
-import { BookingPage } from "@/type/bookings";
+import { booking, BookingPage } from "@/type/bookings";
 
 type BookingStore = {
   bookings: BookingPage;
+  bookingsByUser: booking[];
   loading: boolean;
   fetchBooking: (page: number) => Promise<void>;
   setBooking: (booking: BookingPage) => void;
+  fetchBookingByUser: () => Promise<void>;
+  setBookingByUser: (booking: booking[]) => void;
 };
 
 export const useBookingStore = create<BookingStore>((set) => ({
+  bookingsByUser: [],
   bookings: {
     data: [],
     paging: {
@@ -20,6 +24,7 @@ export const useBookingStore = create<BookingStore>((set) => ({
   },
   loading: true,
   setBooking: (bookings: BookingPage) => set({ bookings }),
+  setBookingByUser: (bookings: booking[]) => set({ bookingsByUser: bookings }),
   fetchBooking: async (page: number = 1) => {
     try {
       const res = await axiosInstance.get("/api/bookings", {
@@ -36,6 +41,18 @@ export const useBookingStore = create<BookingStore>((set) => ({
         },
         loading: false,
       });
+      console.error("Gagal mengambil data field", error);
+    }
+  },
+
+  fetchBookingByUser: async () => {
+    try {
+      const res = await axiosInstance.get(`/api/bookings/users`, {
+        withCredentials: true,
+      });
+      set({ bookingsByUser: res.data, loading: false });
+    } catch (error) {
+      set({ bookingsByUser: [], loading: false });
       console.error("Gagal mengambil data field", error);
     }
   },
